fix(gc): unbreak cargo-test and clear the lint gates for #7803
Three merge blockers, none touching the rooting fixes themselves.

cargo-test (a REQUIRED context) never reached a summary: the branch's
neighborhood dump in `pinned_young_move_report` dereferences
`header_addr - 64 .. +88` with no mapping check, and four pre-existing
tests call it with fabricated addresses (`0x1000 - 64` is guaranteed
unmapped). The binary died with SIGSEGV at test 452 of 2390. A report
printed on the way to an abort must not SIGSEGV and destroy itself, so
both this dump and the native root-slot read now classify against the
arena's page metadata — a real mapping check, not a magnitude guess —
and print a placeholder instead. A stale from-space address, the #7803
case the dump exists for, still classifies into a live space, so the
diagnostic is unchanged where it matters.

2000-line cap: `stack_maps.rs` 2284 -> 1676 and `copying.rs` 2004 ->
1764, by pure code motion. The decode primitives and the object-file
section loaders move to `stack_maps_decode.rs` and
`stack_maps_sections.rs`; the pointer-classification unit moves to
`copying_pointer_set.rs` as a SIBLING of `copying.rs` rather than a
child, because its bodies name `super::gc_moving_loop_polls_enabled`
and `super::malloc::ensure_set_built` — under `gc` those still resolve,
so the move is zero-diff.

addr-class ratchet: `this_binding.rs` used a hand-rolled GcHeader cast.
`addr_class::try_read_gc_header` already exists and is behaviorally
identical here, so route through it rather than take an allowlist
exemption.

Also drops a leftover `// TEMP PROBE` eprintln from the #7803
regression test.

perry-runtime --lib: 2386 passed, 0 failed, 4 ignored.
